Avoid specifing precision explicitly

+++ tacotron/tacotron_v1.py
@@ -54,7 +54,7 @@ class EncoderV1(tf.layers.Layer):
 def AttentionRNNV1(num_units, prenets: Tuple[PreNet],
                    memory, memory_sequence_length):
     rnn_cell = GRUCell(num_units)
-    attention_mechanism = BahdanauAttention(num_units, memory, memory_sequence_length)
+    attention_mechanism = BahdanauAttention(num_units, memory, memory_sequence_length, dtype=memory.dtype)
     return AttentionRNN(rnn_cell, prenets, attention_mechanism)
 
 
@@ -124,7 +124,7 @@ class DecoderV1(tf.layers.Layer):
         decoder_cell = DecoderRNNV1(self.decoder_out_units, attention_cell)
         output_and_done_cell = OutputAndStopTokenWrapper(decoder_cell, self.num_mels * self.outputs_per_step)
 
-        decoder_initial_state = output_and_done_cell.zero_state(batch_size, dtype=tf.float32)
+        decoder_initial_state = output_and_done_cell.zero_state(batch_size, dtype=source.dtype)
 
         helper = TrainingHelper(target,
                                 self.num_mels,
